requirements.txt
beautifulsoup4==4.6.3
Django==2.1.2
djangorestframework==3.9.0
requests==2.19.1
pytest==3.9.3
django-rest-swagger==2.2.0